Revert "Merge branch 'feat/36-search-API-by-date-of-happiness-record' into develop"
This reverts commit 9e5db3f784a2511fbc16751c8290102c7777d426, reversing
changes made to 0c5f1f885aadb4195c4dcd13ff0155f4ff65abf8.
