+++ bonheur/SceneDelegate.swift
@@ -17,7 +17,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window = UIWindow(frame: UIScreen.main.bounds)
         window?.windowScene = windowScene
         
-        self.window?.rootViewController = UINavigationController(rootViewController: LoginViewController())
+        //self.window?.rootViewController = UINavigationController(rootViewController: LoginViewController())
+        self.window?.rootViewController = UINavigationController(rootViewController: CalendarViewController())
         
         window?.makeKeyAndVisible()
     }
