--- bonheur/Sources/NetworkModels/Writing/WritingJSON.swift
@@ -1,13 +1,0 @@
-//
-//  WritingJSON.swift
-//  bonheur
-//
-//  Created by 이정연 on 2023/02/15.
-//
-
-import Foundation
-
-struct WritingJSON: Codable {
-    let contents: String
-    let tagIds: [Int]
-}
