--- bonheur/Sources/NetworkModels/Writing/WritingResponse.swift
@@ -1,12 +1,0 @@
-//
-//  WritingResponse.swift
-//  bonheur
-//
-//  Created by 이정연 on 2023/02/15.
-//
-
-import Foundation
-
-struct WritingResponse: Codable {
-    let boardId: Int
-}
